use Set collection to determine a change exactly, Optimize rendering
